refactor: remove helmut.manager.jobs.Jobs
